% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEIso.R
\name{getDEIso}
\alias{getDEIso}
\title{Differential Isoform Expression Patterns Across Cell Types in Single-Cell RNA-seq Data}
\usage{
getDEIso(
  seurat_obj,
  cluster_column,
  subset_ident,
  gtf,
  min.pct = 0.25,
  rate_cutoff = 0.1,
  multi_cutoff = 0.2,
  ratio_range = 0.05,
  pair_cutoff = c(-1, 0.9),
  ratio_difference = 0.1
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data.
This object should include isoform expression data, and ideally, have cell type annotations or clustering information.
The function will use the expression patterns across cells to identify differentially expressed isoforms
based on their variation across different cell types or clusters.}

\item{cluster_column}{A character string specifying the name of the column in `seurat_obj@meta.data`
that contains cell type annotations or clustering information. This column is used to group cells
based on their cell type or cluster for differential isoform expression analysis.}

\item{subset_ident}{A character vector specifying the cell types or clusters of interest to be extracted
from the `cluster_column` in `seurat_obj@meta.data`. Only the cells corresponding to the specified cell types
or clusters will be included in the differential isoform expression analysis.}

\item{gtf}{A GTF file containing the isoform IDs and corresponding gene IDs.
This file is used to map the isoforms in the single-cell RNA-seq data to their respective genes
for differential expression analysis of isoforms across cell types or clusters.}

\item{min.pct}{A numeric value specifying the minimum percentage of cells in a cluster
that must express an isoform for it to be considered in the differential expression analysis.
This parameter is passed to the `min.pct` argument in Seurat's `FindAllMarkers` function.
Isoforms with expression below this threshold across cell types will be filtered out.
The default value is 0.25 (i.e., isoforms expressed in at least 25% of cells in a cluster).}

\item{rate_cutoff}{A numeric value specifying the minimum ratio of isoform expression to gene expression
required for an isoform to be considered in the differential expression analysis.
Isoforms with a ratio (rate) below this threshold will be filtered out.
The default value is 0.1, meaning isoforms must account for at least 10% of the total gene expression to be included.}

\item{multi_cutoff}{A numeric value specifying the maximum adjusted R-squared value allowed for an isoform
to be considered as showing differential expression patterns across more than two cell types.
The linear regression function used is adjusted R-squared (adj.R.squared).
Isoforms with an adjusted R-squared value greater than this threshold will be filtered out.
The default value is 0.2, meaning only isoforms with an adjusted R-squared value less than 0.2,
indicating weak or less significant expression pattern differences across multiple cell types, will be retained.}

\item{ratio_range}{A numeric value specifying the minimum change in the isoform's expression as a proportion of the total gene expression
across different cell types. This parameter filters isoforms based on the variability in their expression ratio across multiple cell types.
The default value is 0.05, meaning isoforms must show at least a 5% difference in their expression ratio across cell types to be considered.}

\item{pair_cutoff}{A numeric vector of length 2 specifying the range for the product of Pearson correlation coefficient
and cosine similarity used to filter isoforms showing differential expression patterns between two cell types.
Isoforms with a product of these two values outside of this range will be excluded.
The default value is c(-1, 0.9), meaning only isoforms with a product of Pearson correlation coefficient and cosine similarity
greater than -1 and less than 0.9 will be retained.}

\item{ratio_difference}{A numeric value specifying the minimum difference in the isoform's expression ratio as a proportion of the total gene expression
between two cell types. This parameter filters isoforms based on the absolute difference in their expression ratios across the two cell types.
Isoforms with a difference smaller than this threshold will be excluded. The default value is 0.1, meaning isoforms must show at least a 10% difference
in their expression ratios between the two cell types to be considered.}
}
\value{
A list containing four data frames:
1. **multi.gene**: A data frame with isoforms showing differential expression patterns across more than two cell types, relative to gene expression.
2. **multi.major**: A data frame with isoforms showing differential expression patterns across more than two cell types, relative to major isoform expression.
3. **pair.gene**: A data frame with isoforms showing differential expression patterns between two cell types, relative to gene expression.
4. **pair.major**: A data frame with isoforms showing differential expression patterns between two cell types, relative to major isoform expression.

Each data frame contains isoform IDs, gene IDs, and other statistical metrics used for filtering (e.g., correlation values, expression ratios).
Only isoforms that meet the specified thresholds for differential expression patterns across the relevant cell types will be included in the respective data frames.
}
\description{
Differential Isoform Expression Patterns Across Cell Types in Single-Cell RNA-seq Data
}
