% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEIso.R
\name{plotDEIso}
\alias{plotDEIso}
\title{Plot Isoform, Major Isoform, and Gene Expression Across Cell Types}
\usage{
plotDEIso(seurat_obj, gtf, cluster_column, subset_ident, transcript_id)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data.
This object should include isoform expression data, and ideally, have cell type annotations or clustering information.
The function will use the expression patterns across cells to identify differentially expressed isoforms
based on their variation across different cell types or clusters.}

\item{gtf}{A GTF file containing the isoform IDs and corresponding gene IDs.
This file is used to map the isoforms in the single-cell RNA-seq data to their respective genes
for differential expression analysis of isoforms across cell types or clusters.}

\item{cluster_column}{A character string specifying the name of the column in `seurat_obj@meta.data`
that contains cell type annotations or clustering information. This column is used to group cells
based on their cell type or cluster for differential isoform expression analysis.}

\item{subset_ident}{A character vector specifying the cell types or clusters of interest to be extracted
from the `cluster_column` in `seurat_obj@meta.data`. Only the cells corresponding to the specified cell types
or clusters will be included in the differential isoform expression analysis.}

\item{transcript_id}{A character string representing the isoform of interest.
It must be a valid transcript ID present in the GTF file provided.}
}
\value{
A line plot displaying the expression levels of the input `transcript_id` isoform,
  its corresponding major isoform, and the gene across different cell types.
}
\description{
Plot Isoform, Major Isoform, and Gene Expression Across Cell Types
}
